refactor source location on operation dir
